+++ src/Exception/ConfigurationException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace rainwaves\PayfastPayment\Exception;
+
+class ConfigurationException extends PayFastException
+{
+}
+